Rename event_sync to sync
